rename command header
