+++ include/z64quest_hint_commands.h
