+++ src/elf_message/elf_message_ydan.c
@@ -1,5 +1,5 @@
 #include "global.h"
-#include "z64hint_commands.h"
+#include "z64quest_hint_commands.h"
 
 static QuestHintCmd sDungeonNaviQuestHints[] = {
     QUEST_HINT_END(0x5F),
